Initial ML Process Started
Modeling Data - Response to an Offer (Logistic Regression)
We have a better understanding of the demographics and what offers are completed. Let's build a model that predicts whether people will respond to an offer or not.

Problem Definition:
Your task is to combine transaction, demographic and offer data to determine which demographic groups respond best to which offer type.

Evaluation
With this type of classification outcome we need to rely on accuracy to be our indicator of success. If we can reach 90% accuracy at predicting whether or not a person will complete am offer we well pursue the project.

Features
We have to look at what features and target would be best to model the data with. As we are looking to combine demographics and offer types:
Some features to consider
Event
Person
offer_id
amount
reward
offer_type
age
gender
income

Approach:
Combine all data together to get access to the features. Prioritize oferr_type as the target into the model as we are focusing on what offers are completed with the demographic variations as features.